refactor : Tab 컴포넌트에서 ProfileCard로 프로필 정보 전달
상위 컴포넌트에서 하위 컴포넌트로 profile 정보를 props로 전달
